+++ marketplace/core/types/channels/weni_web_chat/views.py
@@ -1,7 +1,9 @@
 from rest_framework import viewsets
+from rest_framework.decorators import action
+from rest_framework.response import Response
 
 from marketplace.applications.models import App
-from .serializers import WeniWebChatSerializer
+from .serializers import WeniWebChatSerializer, WeniWebChatConfigureSerializer
 from . import type as type_
 
 
@@ -16,3 +18,16 @@ class WeniWebChatViewSet(viewsets.ModelViewSet):
 
     def perform_create(self, serializer):
         serializer.save(code=type_.WeniWebChatType.code)
+
+    @action(detail=True, methods=["PATCH"])
+    def configure(self, request, **kwargs):
+        """
+        Adds a config on specified App and create a channel on weni-flows
+        """
+        self.serializer_class = WeniWebChatConfigureSerializer
+        serializer = self.get_serializer(self.get_object(), data=request.data)
+        serializer.is_valid(raise_exception=True)
+
+        self.perform_update(serializer)
+
+        return Response(serializer.data)