+++ CHANGELOG.md
@@ -1,8 +1,15 @@
+v0.1.3
+----------
+ * Add configure endpoint to Weni Web Chat App
+ * Implemente ConfigSerializer on Weni Web Chat Serializer
+ * Add Base64ImageField custom field
+ * Configure celery on project
+ * Add connect client
+
 v0.1.2
 ----------
  * Add platform on create a WeniWebChat app
  * Change WeniWebChatViewSet lookup_field to uuid
- * Add WeniWebChat App create end retrieve endpoint (#36)
  * Set null=True on field App.config
  * Add weni-web-chat app view and configure urls
  * Add WeniWebChat App create end retrieve endpoint